# Cleanup of LFVM interpreter loop