This PR eliminates redundant code from the LFVM interpreter loop. This cuts 4-11% of the execution time for dispatcher-intensive benchmarks like Fib.

```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
                  │ baseline.txt │             modified.txt             │
                  │    sec/op    │    sec/op     vs base                │
Fib/10/lfvm-12       179.4µ ± 5%   160.7µ ±  5%  -10.45% (p=0.002 n=10)
Fib/10/lfvm-si-12    127.5µ ± 7%   116.0µ ± 10%   -9.04% (p=0.004 n=10)
geomean              151.3µ        136.5µ         -9.75%
```

```
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                  │ baseline.txt │            modified.txt             │
                  │    sec/op    │   sec/op     vs base                │
Fib/10/lfvm-16       118.8µ ± 0%   114.0µ ± 2%   -4.08% (p=0.000 n=10)
Fib/10/lfvm-si-16    92.74µ ± 1%   81.93µ ± 0%  -11.66% (p=0.000 n=10)
geomean              105.0µ        96.63µ        -7.94%
```

Pending:
- [x] run CT check